backend/requirements.txt
flask
flask-cors
torch
nemo_toolkit[all]